+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Support for removing or flattening annotations to the `hexapdf modify` command
 - Option to CLI command `hexapdf form` to allow generation of a template file
 - Support for centering a floating box in [HexaPDF::Layout::Frame]
+- [HexaPDF::Type::Catalog#names] for easier access to the name dictionary
 
 ### Changed